# perf: Faster single bracket querying of a graph