@schochastics 2025-01-19
Manipulating a graph via this logic was moved to #1661