@maelle 2025-01-23
[`devtools::test_coverage_active_file()`](https://devtools.r-lib.org/reference/test.html)