@schochastics 2025-01-27
I am now convinced that this is as good as it gets. Any other approach seems to loose too mach performance