@schochastics 2025-01-30
Non unique i/j are now handled outside of `get_adjacency_submatrix()`. I think this is indeed very much cleaner.